initial configuration
user config is fetched from a repo, and it is not activated.

`home-manager` does not specify a nix-store path for `nix-build`, and the environment is not available at `userActivationScript` time, so we cannot do anything here.

we do not want to fetch user config every time as this may fail, and hence fail the complete boot, so we do this only once on `nix-install`. as a result the config is self-modifying, which on its own is scary enough. but more significantly this makes it more complicated to manage machine configuration after bootstrapping.

instead fetching should be done conditionally, e.g. check for file presence beforehand. the more important goal though is to instantiate user config at build time, every time. will investigate patching `home-manager`.