let user service copy config
this does not work as intended yet. while the service description is
copied into $HOME - which indicates that it counts as part of the user
environment - the service itself does not run on login.

maybe the service is still misconfigured. but on the other hand, if we
copy config manually to $HOME and run `home-manager switch`, it is
executed properly. I have to find an example service which is supposed
to start on login and see if it really does after system-build.