add setup for secrets
the general idea is that for initial setup of the system (or user
account) a known storage medium has to be mounted off which we can copy
secrets into a location outside of the store.

a more automated setup would only copy a single gpg keyset with
a human-memorizable password, to use with `gpg-agent` and `ssh` for
fetching all the other secrets off some private (or locally hosted) git
repo. but I first learned of this possibility today and it requires some
work to wire everything together.