use X clipboard for pasting in (u)xterm
why are these kind of things not the default? why would anyone *not*
want to use the clipboard to paste between applications?!

I can only imagine how much pain it takes for a complete beginner to the
subject to create a mental model of what is wrong, phrase the issue in a
way an internet search turns out meaningful hints at the correct
terminology to use for the *actual* search, and then being lucky enough
to find a complete and possibly well-explained solution. and even then,
anyone who got past that stage will have wasted precious life time for
something completely nonsensical.