use xsecurelock for locking screen on sleep
but what I would really like to do and only one person dared to ask
publicly [1] is lock all the user's sessions somehow.

there are only hints at how this could work. someone asked how to run a
script on session-(un)lock [2], and it looks like there is a DBus event
to listen for [3].

[1]: https://superuser.com/questions/965697/tty-and-x-session-locking/965712#965712
[2]: https://unix.stackexchange.com/questions/28181/run-script-on-screen-lock-unlock
[3]: https://unix.stackexchange.com/a/449013