add install script
instead of running `home-manager` explicitly at activation, the system
now just starts an executable which does all the rest.

the trick here is that at bootstrap time the machine-specific
`home-manager` configuration is chosen by the system via the `install`
script, but the environment created in the process contains
a `home-manager` wrapper which uses the same machine-specific
configuration file via the `config.machine` option.

most importantly, the `home-manager` version is the same in both
instances, as it is sourced from the same module. also note that its
release version is now decoupled from the system. while that introduces
double bookkeeping in that regard, it strictly decouples system from user
config, and would be unavoidable for macOS anyway.