re-introduce xmobar
maybe the examples available are too old, but the `manageDocks` hook
does not allow toggling the bar properly. now the config follows [1],
but fixes the problem with the `docks` wrapper I saw somewhere else.

all in all XMonad presents a nightmarish user experience with regard to
configuration. while the compiled, library-type config implements the
suckless ideal to the fullest, it requires learning a major but as of
now still somewhat obscure programming language, and even worse (the
real problem for me right now), having all the development workflow and
tools at hand to get anything out of it. (being able to navigate into
libraries would be really nice.) without `nix` it would probably just as
hard to install the thing in the first place.

I really like the approach of LeftWM [2], but it is so fresh as to be
not yet productively usable.

for now I really need a toggleable status bar with a clock, everything
else is secondary. I tried `polybar` and `tint2`, but both just run and
stay in place - i.e.  `ToggleStruts` simply changes the window offset,
but does not "hide" the bar. I saw one hack [3] which wraps the program
in question into a clumsy process toggle. while one could make it
work more elegantly in `systemd`, this introduces an entirely unneeded
dependency and terrible overhead. `tint2` has such enormous startup
time, that may not even be practical. the bad performance is a pity
since I very much like its flexible yet convenient configuration
paradigm.

the problem is that XMonad is entirely opaque to me, and even if I could
find out how to simply make it overlap the panel with the appropriate
equivalent of `ToggleStruts`, there seems to be a still unsolved bug
which may prevent it to work [4].

eventually it looks like I may come back to `i3`, which I was only
superficially testing at the very beginning. what appalled me was the
unintuitive default configuration and in many respects a lacking
default feature set as compared to XMonad. but more manageable
configuration, more readable documentation and plenty of examples and
plugins seem to be worth the initial hassle of adapting the defaults to
what I am used to from XMonad.

[1]: <https://beginners-guide-to-xmonad.readthedocs.io/configure_xmonadhs.html>
[2]: https://github.com/leftwm/leftwm
[3]: https://bbs.archlinux.org/viewtopic.php?pid=868247#p868247
[4]: https://github.com/xmonad/xmonad-contrib/issues/211