draft for dzen config
as it turns out I was wrong abbout toggling status bars in xmonad.
further inspection of issue #211 [1] revealed that it is only a race
condition, such that restarting xmonad or the bar breaks management
only for existing windows. new windows toggle properly. this means
I can use any implementation I like as long as it advertises the EWMH
dock property. the workaround presented there can still be used in the
config, or I just wait for an upstream fix.

for now I am stuck at the choice between `dzen2` and `conky`.

`dzen2` is essentially a line-based string API for `X`. it is de facto
abandonware. it could use a proper configuration framework just to
escape from the bash hacking madness to produce proper formatting
strings. this is essentially what PyBar [2] advertises to do, although
I did not look into details. most probably I would eventually write my
own little hack in Python.

`conky` has a Lua API and a huge number of built-in features. as it
turns out, it can even act as a EWMH dock via the `own_window_type`
option. with a proper API one can even build sophisticated things
without resorting to madness (as in [3]).

well, I learned a lot (and also much nonsense) today. the main lesson is
probably that writing X applications like a bar-like status display is
not all that hard - but still barely anyone does it, because it is
usually pointless. then again, doing that in a way that is somewhat sane
to configure or extend is much harder, because you inherit all the
problems of API design.

most probably it will be `conky`...

[1]: https://github.com/xmonad/xmonad-contrib/issues/211
[2]: https://github.com/TheUnknownCylon/PyBar
[3]: https://bitbucket.org/mjs7231/conky-pkmeter/src/default/