fancy up `xmobar`
also revert commit d3fe2c28e83a48633beeba35d998540f27577ce1.

the problem with everything else is that setup is simply too
complicated. `tint2` does not work properly in `NixOS`, it does not find
icon sets. `conky` is too much `Lua`, and I found the API too shallow
and clumsy. I would actually like to have proper graphs for system
monitors (not monochromatic as in `conky`). but none of the graphical
bars allow this without programming it almost from scratch.

I took a closer look at `blighty`[1] and probably it is more of what
I want in the long run, just that it is so broad in concept that any
type of configuration is very low-level coding.

for now the bar does everything I need - with options to improve even
more with icon patterns - and even window alignment options are by
default smarter in `xmobar` than anywhere else.

there is just one thing left, namely that the wifi monitor dies on
suspend. the workaround is to manually reload `xmobar` in that case,
which is really fast anyway. this only requires a key mapping for
`xmonad` (see `xmonad.hs`). if the problem persists, it may be more
convenient to run another `systemd` task on wakeup, which would only
restart `xmobar`.

[1]: https://github.com/P403n1x87/blighty