split home-config in check and init
`network-online` was very slow and thus slowed down the also slow
`display-manager` even more. and all this to find out in 99% of all
cases that there is nothing to do. the network requirement here was
guarding against inconsistency on first boot.

now there is an intermediate unit which checks the necessity to fetch
something early on, and only the initialisation unit needs to wait for
`network-online` to be able to fetch remote data if needed.

I first thought the relation here needs to be "init `Requires=,After=`
check", as per `systemd` documentation wording [1] the latter will not start if
the former fails. the fail would happen if the respective git repo
already has the proper remote set.

[1]:
  > If one of the other units fails to activate, and an ordering
  > dependency After= on the failing unit is set, this unit will not be
  > started.

  https://www.freedesktop.org/software/systemd/man/systemd.unit.html#Requires=

sadly this does not work as expected. a failed dependency with
`Requires=,After=` fails this unit, but not lazily (with respect to
requirement satisfaction), i.e. it waits for all other dependencies to
be started before failing this unit.

so now the initialisation unit is triggerd by `OnFailure=`, and for that
to make sense we flip the status code of the relevant check. this is not
really clean, as technically the check script could *actually* fail
somewhere on the way, thereby triggering the initialisation erroneously.
but there is no `OnSuccess=` relation, and I also [2] do not consider
running `systemctl` from the unit script good practice.

maybe one can still hack something together with either some other
relation I have not discovered yet or more clever remapping of error
state.

[2]: https://unix.stackexchange.com/questions/371256/systemd-run-unit-on-success-of-another#comment858486_371342