revert home-config init to "Requires=,After="
seems there was a bug in some intermediate version of the unit
descriptions such that the failing check would not invalidate the
initialisation. `systemd` behaves as documented: if a required
dependency fails, the dependant unit is removed from the schedule
immediately.

but now we have service that fails every time except on first boot.
probably conditional unit execution is more suitable. there exist
`ConditionFirstBoot=` and `ConditionPathExists=`, and I think the path
condition is what I need for the check unit - then the unit execution
can be reduced to `true`.