fix service for fetching password-store
the problem turned out to be a broken derivation for `git-remote-gcrypt`
in NixOS 19.03, which lacked `grep` and `awk`. this was only visible by
replacing the original `git-remote-gcrypt` source with one not
swallowing the output of its internal `git` call. upgrading to 19.09
fixed that.

the other issue surfaced by this detour was that `home-manager` depends
on the system's `nixpkgs`. that kind of makes sense to some degree, but
is right now left implicit. I am not yet convinced that pinning it to
a literal version is the right solution, as it would introduce double
bookkeeping with the system configuration. depending on the system
configuration repository would restore a single source of truth made
explicit, but would also be brittle with respect to that repo changing
its structure and necessarily impure (pinning that repo would not help
to keep them in sync). note again that the whole thing is not portable,
since on darwin there is no system configuration in place to set
`NIX_PATH` appropriately; it would default to some channel version.