activate udiskie service only for GUI machines
this is mostly because home-manager activates this service only for the
systemd `graphical.target`, so it will not be run on my home server
anyway. also the service configuration is very thin, there is not even
a way to declaratively manage the config file. I will have to either
patch the definition or submit a PR.